add state.js, edit AppMain, AppHeader and App.vue
Add dynamic axios call on state.js to generate the call by the searched movie name using the inputs in AppHeader. Removed the axios call in AppMain and import the state.js in all the .vue files.
Generate a dynamic response of the typed film name at the click on the button.
Update readme